# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Form Management API
  description: API for user authentication, form creation, retrieval, and submissions.
  version: 1.0.0

servers:
  - url: http://localhost:5000/api
    description: Local Development Server

paths:
  /auth/signup:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User registered successfully
        "400":
          description: User already exists
        "500":
          description: Server error

  /auth/login:
    post:
      summary: User login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful, returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: Invalid email or password
        "500":
          description: Server error

  /auth/protected:
    get:
      summary: Protected route
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Access granted to protected route
        "401":
          description: Unauthorized

  /forms/save-form:
    post:
      summary: Save a new form
      tags:
        - Forms
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                formName:
                  type: string
                fields:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      type:
                        type: string
      responses:
        "201":
          description: Form saved successfully
        "400":
          description: Missing required fields
        "500":
          description: Error saving form

  /forms:
    get:
      summary: Get all forms for the logged-in user
      tags:
        - Forms
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Returns an array of forms
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    form_id:
                      type: string
                    form_name:
                      type: string
                    fields:
                      type: array
                      items:
                        type: object
                    created_at:
                      type: string
                      format: date-time
        "401":
          description: Unauthorized
        "500":
          description: Error fetching forms

  /forms/all-forms:
    get:
      summary: Get all forms (admin)
      tags:
        - Forms
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Returns all available forms
        "500":
          description: Error fetching all forms

  /forms/fill-form/{formId}:
    get:
      summary: Fetch a specific form by ID
      tags:
        - Forms
      security:
        - bearerAuth: []
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Returns form details
        "404":
          description: Form not found
        "500":
          description: Error fetching form

  /forms/submit:
    post:
      summary: Submit a form response
      tags:
        - Forms
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                formName:
                  type: string
                responses:
                  type: object
                  additionalProperties:
                    type: string
      responses:
        "201":
          description: Form submitted successfully
        "400":
          description: Missing required fields
        "500":
          description: Error submitting form response

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT